[C10D] Support group_dst/group_src in c10d collectives
Addressing RFC 0042 (https://github.com/pytorch/rfcs/pull/71)

It's annoying that 'dst' (for send) ust be a global rank even when a
group is passed in.  But we can't easily change 'dst' without breaking
existing cases.

Furthermore, requiring use of 'global' dst breaks the less common usage
pattern of creating a new ProcessGroup object that is not connected to
the 'default group' and thus has no logical 'global' ranks.

ghstack-source-id: 72264a21bf53bafd0b16b7cbb961aa91cc9b5992
Pull Request resolved: https://github.com/pytorch/pytorch/pull/140460